requirements.txt
win32core==221.23
win32gui==221.3
Pillow==3.2.0
